Renamed ClientStorage as TextClientStorage
