+++ src/Socket/Workerman/TextClientStorage.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace ChessServer\Socket\Ratchet;
+namespace ChessServer\Socket\Workerman;
 
 use ChessServer\Game\PlayMode;
 use ChessServer\Socket\TextClientStorageInterface;
 use Monolog\Logger;
 
-class ClientStorage extends \SplObjectStorage implements TextClientStorageInterface
+class TextClientStorage extends \SplObjectStorage implements TextClientStorageInterface
 {
     private Logger $logger;
 
@@ -24,7 +24,7 @@ class ClientStorage extends \SplObjectStorage implements TextClientStorageInterf
     {
         $this->rewind();
         while ($this->valid()) {
-            if ($id === $this->current()->resourceId) {
+            if ($id === $this->current()->id) {
                 $this->detach($this->current());
             }
             $this->next();
@@ -35,7 +35,7 @@ class ClientStorage extends \SplObjectStorage implements TextClientStorageInterf
     {
         $this->rewind();
         while ($this->valid()) {
-            if ($id === $this->current()->resourceId) {
+            if ($id === $this->current()->id) {
                 $this->current()->send(json_encode($res));
                 $this->logger->info('Sent message', [
                     'id' => $id,
@@ -48,10 +48,11 @@ class ClientStorage extends \SplObjectStorage implements TextClientStorageInterf
 
     public function sendToMany(array $ids, array $res): void
     {
+        $json = json_encode($res);
         $this->rewind();
         while ($this->valid()) {
-            if (in_array($this->current()->resourceId, $ids)) {
-                $this->current()->send(json_encode($res));
+            if (in_array($this->current()->id, $ids)) {
+                $this->current()->send($json);
                 $this->logger->info('Sent message', [
                     'ids' => $ids,
                     'cmd' => array_keys($res),
@@ -63,9 +64,10 @@ class ClientStorage extends \SplObjectStorage implements TextClientStorageInterf
 
     public function sendToAll(array $res): void
     {
+        $json = json_encode($res);
         $this->rewind();
         while ($this->valid()) {
-            $this->current()->send(json_encode($res));
+            $this->current()->send($json);
             $this->next();
         }
     }
